Add storage check
