+++ apps/wallet-mobile/src/features/Notifications/common/storage.ts
@@ -1,40 +1,9 @@
-import {App} from '@yoroi/types'
+import {mountAsyncStorage} from '@yoroi/common'
 
-export const buildProcessedNotificationsStorage = (storage: App.Storage) => {
-  const getValues = async () => {
-    return (await storage.getItem<string[]>('processed')) || []
-  }
+const appStorage = mountAsyncStorage({path: '/'})
 
-  const addValues = async (values: string[]) => {
-    const processed = await getValues()
-    const newProcessed = [...processed, ...values]
-    await storage.setItem('processed', newProcessed)
-  }
+const notificationStorage = appStorage.join('notifications/')
 
-  const setValues = async (values: string[]) => {
-    await storage.setItem('processed', values)
-  }
-
-  const includes = async (value: string) => {
-    const processed = await getValues()
-    return processed.includes(value)
-  }
-
-  const clear = async () => {
-    await storage.setItem('processed', [])
-  }
-
-  const isEmpty = async () => {
-    const processed = await getValues()
-    return processed.length === 0
-  }
-
-  return {
-    getValues,
-    addValues,
-    includes,
-    clear,
-    setValues,
-    isEmpty,
-  }
-}
+export const eventsStorage = notificationStorage.join('events/')
+export const configStorage = notificationStorage.join('settings/')
+export const uiStorage = notificationStorage.join('ui/')
