+++ apps/wallet-mobile/src/features/Notifications/illustrations/PhoneBell.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import Svg, {Path, Defs, LinearGradient, Stop} from 'react-native-svg'
+import Svg, {Defs, LinearGradient, Path, Stop} from 'react-native-svg'
 
 export const PhoneBell = () => {
   return (
@@ -8,54 +8,67 @@ export const PhoneBell = () => {
         d="M150.058 0h-93.35c-4.89 0-9.15 4-9.467 8.89L35.185 195.11C34.868 199.999 38.61 204 43.5 204h93.351c4.89 0 9.15-4.001 9.466-8.891L158.374 8.891C158.69 4 154.949 0 150.058 0z"
         fill="url(#paint0_linear_8_10925)"
       />
+
       <Path
         opacity={0.7}
         d="M147.095 0H56.708c-4.89 0-9.15 4-9.467 8.89L35.185 195.11C34.868 199.999 38.61 204 43.5 204h90.388c4.89 0 9.15-4.001 9.467-8.891L155.411 8.891C155.727 4 151.985 0 147.095 0z"
         fill="url(#paint1_linear_8_10925)"
       />
+
       <Path
         opacity={0.8}
         d="M137.052 190h-94.47c-1.492 0-2.673-1.327-2.577-2.895L50.371 18.544C50.459 17.113 51.586 16 52.948 16h94.469c1.493 0 2.674 1.327 2.577 2.895l-10.365 168.561c-.088 1.431-1.215 2.544-2.577 2.544z"
         fill="#52FCFF"
       />
+
       <Path
         d="M117.379 10.784H81.241c-.796 0-1.4-.645-1.348-1.44a1.56 1.56 0 011.534-1.44h36.139c.795 0 1.399.644 1.347 1.44a1.559 1.559 0 01-1.534 1.44zM89.172 197.32c-.1 1.955-1.767 3.539-3.722 3.539-1.956 0-3.458-1.585-3.357-3.539.1-1.955 1.767-3.54 3.723-3.54 1.955 0 3.458 1.585 3.356 3.54z"
         fill="#fff"
       />
+
       <Path
         opacity={0.2}
         d="M128.642 62.973H43.739l-5.393 83.307h90.296c4.89 0 8.891-4 8.891-8.89V71.863c0-4.89-4.001-8.89-8.891-8.89z"
         fill="#1C5C5D"
       />
+
       <Path
         d="M120.734 138.873H16.682c-4.89 0-8.891-4.001-8.891-8.891V64.455c0-4.89 4-8.891 8.89-8.891h104.054c4.89 0 8.891 4 8.891 8.89v65.528c-.001 4.89-4.002 8.891-8.892 8.891z"
         fill="url(#paint2_linear_8_10925)"
       />
+
       <Path
         d="M4.953 54.608a.792.792 0 01-.562-.232l-1.423-1.424a.793.793 0 111.122-1.123l1.424 1.424a.793.793 0 01-.561 1.355zM2.807 59.592H.794a.794.794 0 110-1.589h2.013a.794.794 0 110 1.589zM10.007 52.46a.793.793 0 01-.791-.737l-.144-2.009a.794.794 0 111.583-.113l.144 2.008a.794.794 0 01-.792.851z"
         fill="#48EBF1"
       />
+
       <Path d="M100 138l19 19v-19h-19z" fill="#407BFF" />
+
       <Path
         d="M86.353 116.316l-1.547-6.744a12.158 12.158 0 01-.013-5.378l2.152-9.607c1.918-8.566-2.429-17.123-9.96-20.889.064-.178.137-.352.18-.542a4.546 4.546 0 00-8.872-1.988c-.042.19-.05.38-.068.568-8.419.192-16.002 6.077-17.92 14.642l-2.152 9.607a12.173 12.173 0 01-2.307 4.858l-4.277 5.44c-1.824 2.32-.603 5.749 2.276 6.394l37.721 8.45c2.88.645 5.447-1.935 4.787-4.811zM57.866 118.399c-1.241-.278-2.356.798-2.142 2.052a6.525 6.525 0 0011.781 2.64c.728-1.043.18-2.493-1.062-2.771l-8.577-1.921zM28.898 105.017c-.384 0-.761-.186-.99-.53-3.78-5.66-5.128-12.45-3.798-19.123 1.33-6.673 5.18-12.429 10.838-16.207a1.191 1.191 0 011.323 1.982 22.94 22.94 0 00-9.825 14.69 22.94 22.94 0 003.442 17.336 1.192 1.192 0 01-.99 1.852z"
         fill="#fff"
       />
+
       <Path
         d="M33.947 101.646c-.384 0-.761-.185-.99-.529a19.276 19.276 0 01-2.893-14.567 19.277 19.277 0 018.255-12.344 1.192 1.192 0 011.323 1.982A16.912 16.912 0 0032.4 87.017a16.913 16.913 0 002.536 12.778 1.19 1.19 0 01-.989 1.851z"
         fill="#fff"
       />
+
       <Path
         d="M38.996 98.276a1.19 1.19 0 01-.99-.53 13.24 13.24 0 01-1.988-10.008 13.245 13.245 0 015.672-8.483 1.192 1.192 0 011.323 1.982 10.876 10.876 0 00-4.658 6.967 10.876 10.876 0 001.632 8.22 1.191 1.191 0 01-.99 1.852zM97.574 121.782a1.192 1.192 0 01-.547-2.249 22.946 22.946 0 0011.407-13.5 22.948 22.948 0 00-1.482-17.612 1.19 1.19 0 112.117-1.093c3.121 6.046 3.702 12.946 1.634 19.428-2.068 6.483-6.537 11.772-12.583 14.894a1.208 1.208 0 01-.546.132z"
         fill="#fff"
       />
+
       <Path
         d="M94.79 116.389a1.191 1.191 0 01-.548-2.25 16.91 16.91 0 008.408-9.951 16.913 16.913 0 00-1.093-12.982 1.192 1.192 0 012.117-1.093 19.276 19.276 0 011.245 14.799 19.281 19.281 0 01-9.584 11.344c-.175.09-.362.133-.546.133z"
         fill="#fff"
       />
+
       <Path
         d="M92.004 110.994a1.191 1.191 0 01-.547-2.249 10.873 10.873 0 005.409-6.402c.888-2.787.64-5.752-.703-8.351a1.191 1.191 0 112.117-1.093c3.373 6.532.802 14.59-5.73 17.964a1.211 1.211 0 01-.546.131z"
         fill="#fff"
       />
+
       <Defs>
         <LinearGradient
           id="paint0_linear_8_10925"
@@ -66,15 +79,24 @@ export const PhoneBell = () => {
           gradientUnits="userSpaceOnUse"
         >
           <Stop stopColor="#1D45BA" />
+
           <Stop offset={0.4204} stopColor="#1D47BA" />
+
           <Stop offset={0.5718} stopColor="#1D4EB9" />
+
           <Stop offset={0.6797} stopColor="#1C59B8" />
+
           <Stop offset={0.767} stopColor="#1B6AB6" />
+
           <Stop offset={0.8417} stopColor="#1A80B3" />
+
           <Stop offset={0.9079} stopColor="#199CB0" />
+
           <Stop offset={0.9663} stopColor="#18BBAD" />
+
           <Stop offset={1} stopColor="#17D1AA" />
         </LinearGradient>
+
         <LinearGradient
           id="paint1_linear_8_10925"
           x1={108.483}
@@ -84,12 +106,18 @@ export const PhoneBell = () => {
           gradientUnits="userSpaceOnUse"
         >
           <Stop stopColor="#17D1AA" />
+
           <Stop offset={0.15} stopColor="#1ACBAF" />
+
           <Stop offset={0.37} stopColor="#21B8BC" />
+
           <Stop offset={0.62} stopColor="#2E9BD3" />
+
           <Stop offset={0.89} stopColor="#3F71F1" />
+
           <Stop offset={1} stopColor="#475FFF" />
         </LinearGradient>
+
         <LinearGradient
           id="paint2_linear_8_10925"
           x1={7.79114}
@@ -99,10 +127,15 @@ export const PhoneBell = () => {
           gradientUnits="userSpaceOnUse"
         >
           <Stop stopColor="#17D1AA" />
+
           <Stop offset={0.15} stopColor="#1ACBAF" />
+
           <Stop offset={0.37} stopColor="#21B8BC" />
+
           <Stop offset={0.62} stopColor="#2E9BD3" />
+
           <Stop offset={0.89} stopColor="#3F71F1" />
+
           <Stop offset={1} stopColor="#475FFF" />
         </LinearGradient>
       </Defs>
